requirements.txt
selenium==4.18.1
python-dotenv==1.0.1
pytz==2024.1

openai==0.28.0
google-generativeai=0.8.3
requests==2.31.0

selenium-wire==5.1.0
webdriver_manager==4.0.1